unify sys and iface files
