+++ internal/sysconf/channels.go
@@ -15,23 +15,18 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package ifaceinfo
-
-import (
-	"fmt"
-	"net"
-	"os"
-	"strings"
-)
+package sysconf
 
 
 
-func State(iface *net.Interface) (string, error) {
-    data, err := os.ReadFile(fmt.Sprintf("/sys/class/net/%s/operstate", iface.Name))
+func Channels2() []int {
+	return []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13}
+}
 
-	if err != nil {
-        return "", err
-    }
 
-	return strings.ToUpper(strings.TrimSpace(string(data))), nil
+func Channels5() []int {
+	return []int{
+	    36,  40,  44,  48,  52,  56,  60,  64,  100, 104, 108, 112, 116, 120, 
+	    124, 128, 132, 136, 140, 144, 149, 153, 157, 161, 165,
+	}
 }
