+++ internal/sysconf/get_gateway_ip.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package ifaceinfo
+package sysconf
 
 import (
 	"fmt"
