+++ internal/sysconf/iface_linktype.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package sysinfo
+package sysconf
 
 import (
 	"fmt"
