+++ internal/sysconf/set_chnl.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package ifconfig
+package sysconf
 
 import (
 	"fmt"
